refactor(users): move hashPassword import from src/utils to shared/auth for better modularity The hashPassword function is now imported from a shared location, which enhances code organization and promotes reusability across different parts of the application. This change helps maintain a cleaner project structure by centralizing shared utilities.
